* Remove README changes * Rename eligibility function * extract rust tests to own file
