+++ common/primitives/src/msa.rs
@@ -300,7 +300,7 @@ pub trait MsaKeyProvider {
 	/// the type to use for looking up keys in storage.
 	type AccountId;
 	/// Returns whether adding `new_key` to `msa_id` would be free
-	fn key_eligible_for_free_addition(
+	fn key_eligible_for_subsidized_addition(
 		old_key: Self::AccountId,
 		new_key: Self::AccountId,
 		msa_id: MessageSourceId,
