+++ pallets/msa/src/tests/mod.rs
@@ -11,3 +11,4 @@ mod retirement_tests;
 mod schema_permission_tests;
 mod signature_registry_tests;
 mod signed_extension_tests;
+mod subsidized_add_key_tests;
